extracts password funcitons to a seperate utils folder for re-use in seeding test and dev data
