+++ api/crud/auth_crud.py
@@ -1,7 +1,7 @@
 from fastapi import HTTPException
 from api.models.user_models import User
 from api.schemas.user_schemas import CreateUserRequest
-import bcrypt
+from api.utils.security_utils.password_utils import hash_password, verify_password
 
 
 def create_user(db, request: CreateUserRequest):
@@ -20,15 +20,3 @@ def create_user(db, request: CreateUserRequest):
     db.refresh(new_user)
 
     return {"username": new_user.username, "user_id": new_user.user_id}
-
-
-def hash_password(password):
-    pwd_bytes = password.encode('utf-8')
-    salt = bcrypt.gensalt()
-    hashed_password = bcrypt.hashpw(password=pwd_bytes, salt=salt)
-    return hashed_password
-
-
-def verify_password(plain_password, hashed_password):
-    password_byte_enc = plain_password.encode('utf-8')
-    return bcrypt.checkpw(password=password_byte_enc, hashed_password=hashed_password)
